test: refactor API tests into separate file
